treewide: redo the C api of pq and separate it from pqcli
it used to be that there was an init function which make a new gbinder client every time for each request. this came to be because of how pqcli which was the first tool i develop worked.
this commit works it and separates the logic. pqcli now works like before, but the C api has an init and cleanup which returns a struct with a gbinder client inside, then we can pass that gbinder client to each function.
now that the system is becoming more used, i also added comments for each method describing the inputs and outputs.
this also includes lots of cleanups and styling changes to be more expandable

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>